% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/beta.r
\name{rbeta_between}
\alias{rbeta_between}
\title{Beta distribution betwen given limits}
\usage{
rbeta_between(n, mode, lower = 0, upper = 1)
}
\arguments{
\item{n}{(required) integer. The number of random numbers to generate}

\item{mode}{(required) numeric. The mode, or most likely value, of the returned distribution.ß}

\item{lower}{(optional) numeric. The minimum value to return}

\item{upper}{(optional) numeric. The maximum value to return}
}
\value{
numeric vector with n elements randomly distributed between lower and upper
}
\description{
Returns a beta distributed vector between \code{lower} and \code{upper}.
}
\examples{
rbeta_between(10, 0.3)
rbeta_between(10, 10, 5, 20)
summary(rbeta_between(1000, 10, 5, 20))
}

