% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/beta.r
\name{rbeta_within}
\alias{rbeta_within}
\title{Beta distribution within a defined tolerance interval}
\usage{
rbeta_within(n, mode, lower, upper, confidence_level = 0.9, sigma = NA)
}
\arguments{
\item{n}{(required) integer. The number of random numbers to generate.}

\item{mode}{(required) numeric. The mode, or most likely value, of the returned distribution.}

\item{lower}{(required) numeric. The lower \code{(1 - confidence_level)/2} tolerance limit.}

\item{upper}{(required) numeric. The upper \code{(1 + confidence_level)/2} tolerance limit.}

\item{confidence_level}{(optional) numeric. The confidence level at which interval \code{lower}, \code{upper} is obtained. To specify sigmas instead, set \code{conf.level = NA}.}

\item{sigma}{(optional) numeric.  If \code{confidence_level = NA}, the number of standard deviations between \code{lower} and \code{upper}.}
}
\value{
numeric vector with n elements randomly distributed so that approximately \code{confidence_level * 100} percent of values will fall between \code{lower} and \code{upper}.
}
\description{
Returns a beta-distributed vector within the tolerance interval defined by \code{lower} and \code{upper}.
}
\details{
\code{conf.level} indicates the confidence level of the interval \code{[lower, upper]}. For example, a high-confidence estimate might be 0.9973, or six standard deviations between \code{lower} and \code{upper}.
}
\examples{
rbeta_within(10, 5, 3, 10)
rbeta_within(10, 5, 3, 10, 0.99)
summary(rbeta_within(1000, 5, 3, 10))
}

