% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/norm.r
\name{rnorm_between}
\alias{rnorm_between}
\title{Normal distribution between given limits}
\usage{
rnorm_between(n, lower = 0, upper = 1)
}
\arguments{
\item{n}{(required) integer. The number of random numbers to generate}

\item{lower}{(optional) numeric. The minimum value to return}

\item{upper}{(optional) numeric. The maximum value to return}
}
\value{
numeric vector with n elements randomly distributed between \code{lower} and \code{upper}. If there is a problem, will return \code{NULL} and print an error.
}
\description{
Returns a normally distributed vector between \code{lower} and \code{upper}.
}
\examples{
rnorm_between(10)
rnorm_between(10, 10, 20)
summary(rnorm_between(10000, 10, 20))
}
