% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/rnorm.r
\name{rnorm_within}
\alias{rnorm_within}
\title{Normal distribution within a defined tolerance interval}
\usage{
rnorm_within(n, lower = 0, upper = 1, confidence_level = 0.9,
  sigma = NA)
}
\arguments{
\item{n}{(required) integer. The number of random numbers to generate.}

\item{lower}{(optional) numeric. The lower \code{(1 - confidence_level)/2} tolerance limit.}

\item{upper}{(optional) numeric. The upper \code{(1 + confidence_level)/2} tolerance limit.}

\item{confidence_level}{(optional) numeric. The confidence level at which interval \code{lower}, \code{upper} is obtained. To specify sigmas instead, set \code{confidence_level = NA}.}

\item{sigma}{(optional) numeric.  If \code{confidence_level = NA}, the number of standard deviations between \code{lower} and \code{upper}.}
}
\value{
numeric vector with n elements randomly distributed so that approximately \code{confidence_level * 100} percent of values will fall between \code{lower} and \code{upper}.
}
\description{
Returns a normally distributed vector within the tolerance interval defined by \code{lower} and \code{upper}.
}
\details{
\code{confidence_level} indicates the confidence level of the interval \code{[lower, upper]}. For example, a high-confidence estimate might be 0.9973, or six standard deviations between \code{lower} and \code{upper}.
}
\examples{
rnorm_within(10)
rnorm_within(10, 10, 20)
summary(rnorm_within(10000, 10, 20))
}

