% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/triangle.r
\name{rtriangle_between}
\alias{rtriangle_between}
\title{Triangle distribution between given limits (wrapper around \code{VGAM::rtriangle()})}
\usage{
rtriangle_between(n, peak = 0.3, lower = 0, upper = 1)
}
\arguments{
\item{n}{(required) integer. The number of random numbers to generate}

\item{peak}{(optional) numeric. The location of the peak (or most likely value) of the triangle distribution.}

\item{lower}{(optional) numeric. The minimum value to return}

\item{upper}{(optional) numeric. The maximum value to return}
}
\value{
numeric vector with n elements randomly distributed between lower and upper with most likely value near \code{peak}.
}
\description{
Returns a triangle-distributed vector between \code{lower} and \code{upper}.
}
\examples{
rtriangle_between(10)
rtriangle_between(10, 10, 5, 20)
summary(rtriangle_between(10000, 10, 5, 20))
}
