% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/triangle.r
\name{rtriangle_within}
\alias{rtriangle_within}
\title{Triangle distribution within a defined tolerance interval}
\usage{
rtriangle_within(n, peak = 0.5, lower = 0, upper = 1,
  confidence_level = 0.9, lower_bound = -Inf, upper_bound = Inf)
}
\arguments{
\item{n}{(required) integer. The number of random numbers to generate.}

\item{peak}{(optional) numeric. The location of the peak (most likely value) of the triangle distribution.}

\item{lower}{(optional) numeric. The lower \code{(1 - confidence_level)/2} tolerance limit.}

\item{upper}{(optional) numeric. The upper \code{(1 + confidence_level)/2} tolerance limit.}

\item{confidence_level}{(optional) numeric. The confidence level at which \code{upper} and \code{lower} are entered.}

\item{lower_bound}{(optional) numeric. The lower boundary for returned values;  no random values can be below this value. e.g. for processes where negative values do not make sense, \code{lower = 0}.}

\item{upper_bound}{(optional) numeric. The upper boundary for returned values;  no random values can be above this value.}
}
\value{
numeric vector with n elements randomly distributed so that approximately \code{confidence_level} percent of values will fall between \code{lower} and \code{upper}. If an error occurs, will usually return \code{NULL} and print an error message.
}
\description{
Returns a triangle-distributed vector within the \code{confidence_level} tolerance interval defined by \code{lower} and \code{upper}.
}
\details{
\code{confidence_level} indicates the confidence level of the interval \code{[lower, upper]}. For example, a high-confidence estimate might be 0.9973, or six standard deviations between \code{lower} and \code{upper}.
}
\examples{
rtriangle_within(10)
rtriangle_within(10, 12, 10, 20)
summary(rtriangle_within(10000, 12, 10, 20))
}
