% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unif.r
\name{runif_digits}
\alias{runif_digits}
\title{Uniform distribution integers of fixed length}
\usage{
runif_digits(n, digits = 4)
}
\arguments{
\item{n}{(required) integer. The number of random values to generate.}

\item{digits}{(required) integer. The number of digits desired to the left of the decimal.}
}
\value{
numeric vector with n integer elements uniformly randomly distributed with \code{digits} number of digits. If an error occurs, will usually return \code{NULL} and print an error message.
}
\description{
Returns a uniformly distributed integer vector, where every element has \code{digits} number of digits to the left of the decimal.
}
\examples{
 runif_digits(1, 6)
 runif_digits(10, 6)
}
