% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unif.r
\name{runif_between}
\alias{runif_between}
\title{Uniform distribution between specified limits (wrapper around \code{runif()})}
\usage{
runif_between(n, lower = 0, upper = 1, ...)
}
\arguments{
\item{n}{(required) integer. The number of random numbers to generate.}

\item{lower}{(optional) integer. The minimum value to return.}

\item{upper}{(optional) integer. The maximum value to return.}
}
\value{
numeric vector with n elements randomly distributed between \code{lower} and \code{upper}. If an error occurs, will usually return \code{NULL} and print an error message.
}
\description{
Returns a uniformly distributed vector between \code{lower} and \code{upper}. Bounded as \code{[lower, upper]}.
}
\examples{
 runif_between(10)
 runif_between(10, 10, 20)
 summary(runif_between(10000, 10, 20))
}
